+++ src/app/Components/ComponentsSamples/BottomSheets/Sheets/BottomSheetWithToolbar.xaml.cs
@@ -1,3 +1,4 @@
+using System.Windows.Input;
 using DIPS.Mobile.UI.Components.BottomSheets;
 
 namespace Components.ComponentsSamples.BottomSheets.Sheets;
@@ -7,17 +8,5 @@ public partial class BottomSheetWithToolbar
     public BottomSheetWithToolbar()
     {
         InitializeComponent();
-    }
-
-    protected override void OnClosed()
-    {
-        base.OnClosed();
-        
-        
-    }
-
-    private void MenuItem_OnClicked(object sender, EventArgs e)
-    {
-        BottomSheetService.CloseCurrentBottomSheet();
     }
 }