+++ src/library/DIPS.Mobile.UI/Components/Buttons/Button.cs
@@ -5,11 +5,9 @@ namespace DIPS.Mobile.UI.Components.Buttons
         public Button()
         {
             this.SetAppThemeColor(BackgroundColorProperty, ColorName.color_primary_90);
-            this.SetAppThemeColor(TextColorProperty, ColorName.color_neutral_05);
+            this.SetAppThemeColor(TextColorProperty, ColorName.color_system_white);
             Padding = new Thickness(Sizes.GetSize(SizeName.size_3), Sizes.GetSize(SizeName.size_1));
             CornerRadius = Sizes.GetSize(SizeName.size_2);
         }
-        
-        
     }
 }