+++ src/library/DIPS.Mobile.UI/Components/Chips/iOS/ChipHandler.cs
@@ -3,6 +3,7 @@ using DIPS.Mobile.UI.Resources.Colors;
 using Microsoft.Maui.Handlers;
 using Microsoft.Maui.Platform;
 using UIKit;
+using Color = Microsoft.Maui.Graphics.Color;
 using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
 
 // ReSharper disable once CheckNamespace
@@ -26,7 +27,7 @@ public partial class ChipHandler : ViewHandler<Chip, UIButton>
         
         // Here we style the button as close as possible to native compact datepicker in iOS
         // We do not use the design system here so this does not diverge at a later point
-        Button.Background = new Microsoft.Maui.Graphics.Color(118, 118, 128, 30);
+        Button.Background = Color.FromArgb(BackgroundColorAsHex);
         Button.FontSize = 17;
         Button.CornerRadius = 6;
         Button.Padding = new Thickness(12, 6, 12, 6);