+++ src/tests/unittests/Converters/ValueConverters/DateConverterTests.cs
@@ -40,7 +40,7 @@ namespace DIPS.Mobile.UI.UnitTests.Converters.ValueConverters
         public static IEnumerable<object[]> TestDataForShortFormat =>
             new List<object[]>()
             {
-                new object[] {"no", new DateTime(1991, 12, 12), "12. des 1991"},
+                new object[] {"nb", new DateTime(1991, 12, 12), "12. des 1991"},
                 new object[] {"en-gb", new DateTime(1991, 12, 12), "12th Dec 1991"},
                 new object[] {"en-us", new DateTime(1991, 12, 12), "Dec 12th, 1991"},
             };
@@ -91,10 +91,10 @@ namespace DIPS.Mobile.UI.UnitTests.Converters.ValueConverters
                 new object[] { "en-gb", new DateTime(1990, 12, 12).AddDays(-1), "Yesterday"},
                 new object[] { "en-gb", new DateTime(1990, 12, 12).AddDays(1), "Tomorrow"},
                 new object[] { "en-gb", new DateTime(1990, 12, 10), "10th Dec"},
-                new object[] {"no", new DateTime(1990, 12, 12), "I dag"},
-                new object[] {"no", new DateTime(1990, 12, 12).AddDays(-1), "I går"},
-                new object[] {"no", new DateTime(1990, 12, 12).AddDays(1), "I morgen"},
-                new object[] {"no", new DateTime(1990, 12, 10), "10. des"}
+                new object[] {"nb", new DateTime(1990, 12, 12), "I dag"},
+                new object[] {"nb", new DateTime(1990, 12, 12).AddDays(-1), "I går"},
+                new object[] {"nb", new DateTime(1990, 12, 12).AddDays(1), "I morgen"},
+                new object[] {"nb", new DateTime(1990, 12, 10), "10. des"}
             };
 
         [Theory]